add ChatGPTManager class
